Move StatusText to tree
