+++ modules/study/src/main/GameToRoot.scala
@@ -20,6 +20,6 @@ private final class GameToRoot:
   private def endComment(game: Game) =
     game.finished.option:
       val result = Outcome.showResult(Outcome(game.winnerColor).some)
-      val status = lila.core.game.StatusText.apply(game.status, game.winnerColor, game.variant)
+      val status = lila.tree.StatusText(game.status, game.winnerColor, game.variant)
       val text   = s"$result $status"
       Comment(Comment.Id.make, Comment.Text(text), Comment.Author.Lichess)
