+++ modules/study/src/main/StudyPgnImport.scala
@@ -46,7 +46,7 @@ final class StudyPgnImport:
                 status = res.status,
                 outcome = outcome,
                 resultText = chess.Outcome.showResult(outcome.some),
-                statusText = lila.core.game.StatusText.apply(res.status, res.winner, game.board.variant)
+                statusText = lila.tree.StatusText(res.status, res.winner, game.board.variant)
               )
 
             val commented =
